Get rid of IWatcher.
