+++ sdk/storage/Azure.Storage.Webjobs.Extensions.Blobs/src/Bindings/NotifyingBlobStream.cs
@@ -1,0 +1,38 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System.Threading;
+using System.Threading.Tasks;
+using System.IO;
+
+namespace Microsoft.Azure.WebJobs.Host.Blobs.Bindings
+{
+    internal class NotifyingBlobStream : DelegatingStream
+    {
+        private readonly IBlobCommitedAction _committedAction;
+
+        public NotifyingBlobStream(Stream inner, IBlobCommitedAction committedAction)
+            : base(inner)
+        {
+            _committedAction = committedAction;
+        }
+
+        public override void Flush()
+        {
+            base.Flush();
+            if (_committedAction != null)
+            {
+                _committedAction.Execute();
+            }
+        }
+
+        public override async Task FlushAsync(CancellationToken cancellationToken)
+        {
+            await base.FlushAsync(cancellationToken).ConfigureAwait(false);
+            if (_committedAction != null)
+            {
+                await _committedAction.ExecuteAsync(cancellationToken).ConfigureAwait(false);
+            }
+        }
+    }
+}
